Move `VerifyServiceInstalledAndMatchesClient` to separate file
