+++ lib/vnet/service_windows.go
@@ -19,7 +19,6 @@ package vnet
 import (
 	"context"
 	"log/slog"
-	"os"
 	"syscall"
 	"time"
 
@@ -155,52 +154,6 @@ func (w *handler) Execute(ctx context.Context, args []string) error {
 	}
 	if err := runWindowsAdminProcess(ctx, &cfg); err != nil {
 		return trace.Wrap(err, "running admin process")
-	}
-	return nil
-}
-
-// VerifyServiceInstalledAndMatchesClient returns nil if the service is installed and matches the client version.
-func VerifyServiceInstalledAndMatchesClient() error {
-	// Avoid [mgr.Connect] because it requests elevated permissions.
-	scManager, err := windows.OpenSCManager(nil /*machine*/, nil /*database*/, windows.SC_MANAGER_CONNECT)
-	if err != nil {
-		return trace.Wrap(err, "opening Windows service manager")
-	}
-	defer windows.CloseServiceHandle(scManager)
-	serviceNamePtr, err := syscall.UTF16PtrFromString(serviceName)
-	if err != nil {
-		return trace.Wrap(err, "converting service name to UTF16")
-	}
-	serviceHandle, err := windows.OpenService(scManager, serviceNamePtr, windows.SERVICE_QUERY_CONFIG)
-	if err != nil {
-		return trace.Wrap(err, "opening Windows service %v", serviceName)
-	}
-	service := &mgr.Service{
-		Name:   serviceName,
-		Handle: serviceHandle,
-	}
-	defer service.Close()
-
-	config, err := service.Config()
-	if err != nil {
-		return trace.Wrap(err, "getting service config")
-	}
-	exe, err := os.Executable()
-	if err != nil {
-		return trace.Wrap(err, "getting executable path")
-	}
-	serviceArgs, err := windows.DecomposeCommandLine(config.BinaryPathName)
-	if err != nil {
-		return trace.Wrap(err, "parsing Windows service binary command line")
-	}
-	if len(serviceArgs) == 0 {
-		return trace.BadParameter("Windows service has empty binary command line")
-	}
-
-	// Require exact binary match between the client and service.
-	// Hash comparison is enough here because same binary means same version.
-	if err = compareFiles(exe, serviceArgs[0]); err != nil {
-		return trace.Wrap(err, "comparing tsh.exe executable with service executable")
 	}
 	return nil
 }
